Biological measurements and metadata associated with fish dissected in the lab
Source: R/data.R

A dataset containing variables collected from laboratory dissections of fish caught in the field by the Juvenile Salmon Program. A subset of fish_field_data

fish_lab_data

Format

An object of class tbl_df (inherits from tbl, data.frame) with 5602 rows and 19 columns.

Details

ufn

The alphanumeric string that uniquely identifies a fish. Primary key to: fish_field_data, sealice_field, and sealice_lab_motiles.

semsp_id

A unique ID for every fish retained that was a concatenation of survey_date, site_id, species, and order the fish was pulled out of the net. In 2017 this was method was abandonded, and instead we assigned seine IDs and UFNs to uniquely identify seines and fish

date_processed

The date the fish was processed (dissected) in the lab. Blank infers the fish has not be dissected yet. Format: yyyy-mm-dd

cwt

Fish were scanned using a handheld CWT read. none = no tag detected. cwt = coded wire tag detected

adipose

none = adipose is not clipped. AD = adipose fin was clipped

weight

weight of the fish measured in the lab using Pro Scout scales. Units: grams. 0.1 g precision

standard_length

The length of the fish measured from the tip of the snout to the caudal peduncle where scales give way to caudal fin rays. Units: millimetres

fork_length

The length of the fish meadured from the tip of the snout to the fork of the caudal fin. Units: millimetres

photo_number

Photos of abnormal looking fish were taken. The name of the photo file in the photo database. The file name is also the UFN of the fish.

comments_protocol

Comments related to deviations to standard dissection protocols

comments_fish_health_lab

Comments to describe any abnormal/unique observations about the fish itself (e.g. lesions, non-louse parasites, tissue discolouration, scale loss, etc.)

dissector

The initials of the technician who performed the dissection

dissection_protocol

Categories used to describe the general lab dissection workflow performed on a fish:

Full work-up 1 = procedure for 2015 & 2016 samples (One RNA sample each taken from muscle, gill, brain, spleen, liver, heart, and head kidney; fatty acid, stable isotope, DNA (if no fin clips were taken in the field), stomach, otolith, and scales).

Full work-up 2 = procedure for all samples effective Oct. 2017 (One RNA sample each taken from muscle, gill, brain, spleen, liver, heart, and head kidney; fatty acid, stable isotope, extra muscle stored at -80 degrees Celcius, DNA (if no fin clips were taken in the field), stomach, otolith, and scales).

Full work-up 3 = procedure for fish subsampled for compound specific stable isotope analysis (by David Costalago, Feb. 2018). All samples taken as in full work-up 2, with the addition of a second isotope sample taken.

Work-up excluding cold samples = full work-up 1 without RNA, fatty acid, or stable isotope samples taken.

Irregular work-up = processing with at least one sample type not taken (e.g. no RNA, no DNA, etc.).

Lice enumeration only - in lab = fish was inspected for sea lice only and did not undergo dissection. Lengths and weight recorded.

-20 work up 1 = procedure for dissecting fish that have been previously stored at -20 before dissection. No RNA samples, no fatty acid samples, no extra muscle sample, no sea-lice retention. Reduced sterilization between fish to scrubbing and ethanoling direct contact surfaces with 95

fish_lab_qc_data

The initials of the technician who quality checked the data

fish_lab_qc_flag

yes = one or more columns has an unresolvable error in data entry (not pertaining to the physical quality of the sample). no = no issues with data quality

fish_lab_flag_reason

Explains why the fish has been flagged

fish_lab_qc_log

Describes any action taken by QC to address and resolve flagged issues.

lice_collection_protocol

The protocol that was followed to collect sealice from the fish. Categories:

lab_motiles_cryo = only motile sea lice were collected

lab_fine_2015 = all life stages, sexes, and species of sea lice were collected (only in 2015)

not_collected = no sea lice were collected

lice_id_protocol

The protocol that was followed for identifying sea lice in the lab. Categories:

lportner_fine = fine-scale identification of all sea lice and life stages that were obtained under

lab_motiles = identification of only motile stages of sea lice