Identification of sea lice collected from fish in the lab in 2015 from lice collection protocol *lab_fine_2015*
Source: R/data.R

A dataset containing detailed sea lice taxonomic identification

ufn

The alphanumeric string that uniquely identifies a fish. Foreign key to: fish_lab_data.

date_liced

The date the lice were identified.

lab_count_fine_no_id

The number of lice collected that the dissecting technician recorded when the lice were originally collected

lab_count_fine_total

The total number of lice that the louse identifier counted in the sample vial

lep_cop

Lepeophtheirus salmonis copepodite.

lep_cunifer_cop

Lepeophtheirus cunifer copepodite.

lep_chal_a

Lepeophtheirus salmonis chalimus A.

lep_chal_b

Lepeophtheirus salmonis chalimus B.

lep_pa_m_1

Lepeophtheirus salmonis pre-adult male stage 1.

lep_pa_m_2

Lepeophtheirus salmonis pre-adult male stage 2.

lep_pa_f_1

Lepeophtheirus salmonis pre-adult female stage 1.

lep_pa_f_2

Lepeophtheirus salmonis pre-adult female stage 2.

lep_pa_unid

Lepeophtheirus salmonis pre-adult unidentified stage of sex.

lep_a_m

Lepeophtheirus salmonis adult male.

lep_a_f

Lepeophtheirus salmonis adult female.

lep_grav_f

Lepeophtheirus salmonis gravid female.

cal_cop

Caligus clemensi copepodite.

cal_chal_a_1

Caligus clemensi chalimus A stage 1.

cal_chal_a_2

Caligus clemensi chalimus A stage 2.

cal_chal_b_3

Caligus clemensi chalimus B stage 3.

cal_chal_b_4_f

Caligus clemensi chalimus B stage 4 female.

cal_chal_b_4_m

Caligus clemensi chalimus B stage 4 male.

cal_chal_4_unid

Caligus clemensi chalimus B stage 4 unidentified sex.

cal_chal_a_unid

Caligus clemensi chalimus A unidentified sex.

cal_chal_b_unid

Caligus clemensi chalimus B unidentified sex.

cal_pa_m

Caligus clemensi pre-adult male.

cal_pa_f

Caligus clemensi pre-adult female.

cal_grav_f

Caligus clemensi gravid female.

cal_mot_unid

Motile Caligus clemensi.

unid_louse

Unidentified louse

sealice_comments

Comments pertaining to sealice ID

sealice_lab_finescale

Format

An object of class tbl_df (inherits from tbl, data.frame) with 1486 rows and 33 columns.