Site, date, and other metadata associated with a survey
Source: R/data.R

A dataset containing variables collected from juvenile salmon program site surveys

survey_data

Format

An object of class tbl_df (inherits from tbl, data.frame) with 542 rows and 22 columns.

Details

ufn

Universal Fish Number. For 2015 and 2016 fish, assigned to each specimen in the order they were processed in the lab. For 2017 fish, ufn was assigned in the order they were processed in the field.

survey_id

Primary key to: seine_data, ysi, ctd, zoop_tows. A unique survey ID assigned to every survey

survey_date

yyy-mm-dd. Date upon which a survey was conducted

site_id

unique identifier given for the geographic location of a survey

region

DI = Discovery Islands, JS = Johnstone Strait

zone

Ctegorization of the generalized migration route taken by outmigrating juvenile salmon (West, Central, or East for Discovery Islands; N or S for Johnstone Strait).

site_name

Chart name for the location of a survey site

crew

Full names of crew members

precip

Precipitation. none, light, moderate, or heavy. Subjective categorical representation of the amount of precipitation

cloud_cover

0, 25, 50, 75, or 100 Visual approximation of the amount of cloud cover overhead

sea_state

0, 1, 2, 3, 4, or 5. 0 = Sea-surface smooth and mirror like. 1 = Scaly ripples, no foam crests. 2 = Small wavelets, crests glassy, no breaking. 3 = Large wavelets, crests begin to break, scattered whitecaps. 4 = small waves 1–4 ft. becoming longer, numerous whiteaps. 5 = Moderate waves 4–8 ft. taking longer form, many whitecaps, some spray.

wind_speed

knots. 0, 0.1-5, 5-10, 10-15, 15-20, or 20-30. Estimated wind speed range observed on site

wind_direction

N, S, E, W, NE, NW, SE, or SW. Approximation of which way the wind is coming from

tide_state

high slack, low slack, flood, or ebb tide. Tidal state at the start of the survey

survey_time_start

hh:mm:ss 24-hour time of commencing visual survey for site activity

survey_time_end

hh:mm:ss 24-hour time of end of visual survey for site activity

ctd_cast_id

This is the primary key to join survey data to CTD data. It is an automatically-generated concatenation of Site Code, Date, and drop number

zoop_sample_id

Zooplankton Sample ID. The unique ID that identifies a zooplankton sample that is stored in a 250 mL PET jar

secchi

Metres. Secchi depth reading in metres (rounded to nearest 0.1 m)

net_sets

number of times the seine was deployed in a survey

set_type

This is a category for how or who the survey was conducted. Standard surveys can be used for the purpose of CPUE calculations, other categories should not be

survey_comments

field notes specific to survey conditions

sampling_week

A calculated value. Julian day (j) + 4 days for historical categorization purposes, integer divide by 7. (j + 4 / 7)