Site, date, and other metadata associated with a survey
Source: R/data.R

A dataset containing variables collected from Juvenile Salmon Program site surveys

survey_data

Format

An object of class tbl_df (inherits from tbl, data.frame) with 616 rows and 23 columns.

Details

survey_id

A unique alphanumeric code assigned to every survey. Primary key to: seine_data, site_activity, ysi, ctd_casts, and zoop_tows.

survey_date

Date upon which a survey was conducted. Format: yyyy-mm-dd

site_id

A unique identifier given for the geographic location of a survey

region

General geographic zone of a site. DI = Discovery Islands, JS = Johnstone Strait

zone

Categorization of the generalized migration route taken by outmigrating juvenile salmon (West, Central, or East for Discovery Islands; N or S for Johnstone Strait).

site_name

Chart name for the location of a survey site

crew

Full names of crew members

precip

Subjective categorical representation of the amount of precipitation. Categories: none, light, moderate, heavy

cloud_cover

Visual approximation of the percent of cloud cover overhead. Categories: 0, 25, 50, 75, 100

sea_state

Categories describing surface conditions:

0 = Sea-surface smooth and mirror like.

1 = Scaly ripples, no foam crests.

2 = Small wavelets, crests glassy, no breaking.

3 = Large wavelets, crests begin to break, scattered whitecaps.

4 = small waves 1–4 ft. becoming longer, numerous whiteaps.

5 = Moderate waves 4–8 ft. taking longer form, many whitecaps, some spray.

wind_speed

Estimated wind speed range observed on site. Units: knots. Categories: 0, 0.1-5, 5-10, 10-15, 15-20, 20-30.

wind_direction

Approximation of which way the wind is coming from. Categories: N, S, E, W, NE, NW, SE, SW.

tide_state

Tidal state at the start of the survey. Categories: high slack, low slack, flood, ebb tide.

survey_time_start

24-hour time of commencing visual survey for site activity. Format: hh:mm:ss

survey_time_end

24-hour time of end of visual survey for site activity. Format: hh:mm:ss

ctd_cast_id

An automatically-generated concatenation of site_id, survey_date, and drop number. Primary key to join survey_data to CTD data.

zoop_sample_id

The unique code that identifies a zooplankton sample that is stored in a 250 mL PET jar. Primary key to join survey_data to zoop_tows.

secchi

Secchi depth reading. Units: metres (rounded to nearest 0.1 m)

net_sets

The number of times the seine was deployed in a survey

set_type

How (or for whom) the survey was conducted. Standard surveys can be used for the purpose of CPUE calculations; other categories (non-standard or DFO should be excluded.

survey_comments

Field notes specific to survey conditions

sampling_week

A calculated value. Julian day (j) + 4 days for historical categorization purposes, integer divide by 7. \(j + 4 / 7\)